src/app.py
"""
SparkPreCheckOptimizer - Main Streamlit Application
A UI app that analyzes Azure Data Lake files (Parquet) or Synapse tables and a Spark program,
suggesting configs, checks, and more using Azure OpenAI.
"""

import streamlit as st
import os
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Environment variable placeholders (set these in your .env file)
# AZURE_OPENAI_API_KEY=your_openai_api_key_here
# AZURE_OPENAI_ENDPOINT=your_openai_endpoint_here
# AZURE_STORAGE_ACCOUNT_NAME=your_storage_account_name_here
# AZURE_TENANT_ID=your_tenant_id_here
# AZURE_CLIENT_ID=your_client_id_here
# AZURE_CLIENT_SECRET=your_client_secret_here

def main():
    """Main Streamlit application."""
    st.set_page_config(
        page_title="SparkPreCheckOptimizer",
        page_icon="⚡",
        layout="wide",
        initial_sidebar_state="expanded"
    )
    
    st.title("⚡ SparkPreCheckOptimizer")
    st.markdown("""
    Analyze Azure Data Lake files (Parquet) or Synapse tables and Spark programs to suggest 
    optimal configurations, performance checks, and improvements using Azure OpenAI.
    """)
    
    # Sidebar for configuration
    with st.sidebar:
        st.header("Configuration")
        
        # Data source selection
        data_source = st.selectbox(
            "Data Source",
            ["Azure Data Lake (Parquet)", "Azure Synapse Table", "Upload File"]
        )
        
        # Spark program input
        spark_program = st.text_area(
            "Spark Program (Paste your Spark code here)",
            height=200,
            placeholder="// Paste your Spark program here..."
        )
        
        # Analysis options
        st.header("Analysis Options")
        include_configs = st.checkbox("Suggest Spark Configurations", value=True)
        include_checks = st.checkbox("Performance Checks", value=True)
        include_optimizations = st.checkbox("Optimization Suggestions", value=True)
    
    # Main content area
    if st.button("🚀 Analyze and Optimize", type="primary"):
        if not spark_program.strip():
            st.error("Please provide a Spark program to analyze.")
            return
        
        st.info("Analysis in progress... This feature will be implemented in the next iteration.")
        
        # Placeholder for analysis results
        st.subheader("Analysis Results")
        st.write("This section will display optimization suggestions, configurations, and checks.")
        
        # Placeholder tabs for different analysis sections
        tab1, tab2, tab3 = st.tabs(["📊 Configurations", "🔍 Performance Checks", "⚡ Optimizations"])
        
        with tab1:
            st.write("Spark configuration suggestions will appear here.")
            
        with tab2:
            st.write("Performance analysis and checks will appear here.")
            
        with tab3:
            st.write("Optimization recommendations will appear here.")

if __name__ == "__main__":
    main()
