requirements.txt
streamlit>=1.28.0
pyspark>=3.4.0
openai>=1.0.0
langchain>=0.1.0
langchain-openai>=0.1.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
azure-synapse-spark>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
